# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Api",
        "description": "Demo Car Rental Api",
        "version": "1.0"
    },
    "paths": {
...
        "/car-rental/api/v1/users/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Authenticate user",
                "operationId": "544cd48f8c888ab186fbc18440dc16fe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "johndoe123",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe123"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "10293182301230123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car-rental/api/v1/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "operationId": "69e4959835b07b7c26dbb5842937e2f7",
                "parameters": [
                    {
                        "parameter": "AuthorizationHeader",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer <token>",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "firstname": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "lastname": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "johndoe123"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "operationId": "ce3505aeac8149c112c6c64c4c6c42c4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "firstname": {
                                                "type": "string"
                                            },
                                            "lastname": {
                                                "type": "string"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "firstname": "John",
                                    "lastname": "Doe",
                                    "email": "johndoe123",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Registration successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "User already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/car-rental/api/v1/users/deletion/{id}": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "operationId": "86dba5bab4b5c9a30885077af1f60a99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "parameter": "AuthorizationHeader",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer <token>",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete users success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}